[INTERNAL] Ensure external listerners are always the first the advertised.listeners configuration
This is needed for old clients connecting to kafka through Zookeeper that does not have a way
to infer the right listener.
In this case, the first listener in the advertised.listener config is used to connect to brokers.
This patch ensures the external listeners (those reachable from outside) are listed before internal ones