Use local broker as bootstrap for cruisecontrol metrics reporter
cruise-control metrics reporter is now configured
to use local broker as a bootstrap server.
This is needed to avoid a race condition during initial
cluster creation when the first broker is initialised but its dns
entry in internal K8s DNS is not yet available immediately for the cruise-control reporter to use.